Fix blank first column in multicolumn SimpleMenu
In SmeBSB.c Redisplay, an XClearArea in a libXaw3dXft-specific block failed
to add the rectangle's x origin to the range of x coordinates being cleared.
This caused the first column of a multicolumn SimpleMenu to be blank as the
drawing of the second and subsequent columns would clear the first.