Replace _XawImWcLookupString with libXaw version
src/XawIm.c: In function ‘_XawImWcLookupString’:
src/XawIm.c:1553:15: warning: ‘p’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1553 |         ret = Xutf8LookupString (p->xic, event, tmp_buf, 64, keysym_return, status_return);
      |               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

That line could not possibly work and is dead code.

_XawImWcLookupString is a "private" function declared in XawImP.h.
Its only use is in TextAction.c InsertChar.  The call
happens only if XAW_INTERNATIONALIZATION is defined and _Xaw3dXft->encoding
!= -1.  The broken line is conditional on _Xaw3dXft->encoding == -1.

Comparing the function with its counterpart in libXaw reveals a second
problem:  XwcLookupString is being given the number of bytes when it expects
the number of wchars.

The status_return parameter is not there in libXaw, is never looked at, and
is unnecessary.