Stop pretending to use XComposeStatus
"If present (non-NULL), the XComposeStatus structure records the state, which
is private to Xlib, that needs preservation across calls to XLookupString to
implement compose processing.  The creation of XComposeStatus structures is
implementation-dependent; a portable program must pass NULL for this
argument."

The source of libX11 reveals that the parameter does nothing.  libXaw got rid
of it in XFree86 4.0:

#2544 12 Mar 1999
TextAction.c:
 + Does not call XLookpupString in InsertChar any more, but the new function
   _XawLookupString (in XawIm.c) to work correclty with composed characters.
XawIm.c:
 + Added the private function _XawLookupString, that just calls XmbLookupString,
   or in case of any initialization error, XLookupString.

Every call to X*LookupString subsequently passed NULL for the last argument.